Classical_computing/Computer_vision/Profession classyfication/requirements.txt
numpy>=1.20.0
pandas>=1.3.0 
matplotlib>=3.4.0 
pillow>=8.2.0 
torch>=1.9.0 
torchvision>=0.10.0 
tqdm>=4.62.0 
scipy>=1.7.0 
scikit-learn>=0.24.2 
pytest>=6.2.5 
d2l>=1.0.0 
jupyterlab>=3.1.0 
ipywidgets>=7.6.0 
optuna>=3.0.0 # For hyperparameter optimization 
seaborn>=0.11.2 # For visualization 
tensorboard>=2.10.0 # For training visualization 
torchsummary>=1.5.1 # For model summary 
opencv-python>=4.5.0 # For image processing 
timm>=0.6.12 # For additional computer vision models 
einops>=0.5.0 # For Vision Transformer implementation

